+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.0
 
+- Postprocess the value for a column, e.g. dividing by the number of elements, using `Column::postprocess`.
 - Extract specific Criterion performance statistics, using `ColumnStat`.
 
 # 0.3